+++ TurtleAPI/PolygonIO/PolygonPreviousCloseResult.cs
@@ -7,11 +7,11 @@ namespace TurtleAPI.PolygonIO
     internal class PolygonPreviousCloseResult
     {
         public string T { get; set; } = string.Empty; //Ticker
-        public decimal o { get; set; } //open
-        public decimal c { get; set; } //close
-        public decimal h { get; set; } //high
-        public decimal l { get; set; } //low
-        public decimal n { get; set; } // (!)investigate
+        public double o { get; set; } //open
+        public double c { get; set; } //close
+        public double h { get; set; } //high
+        public double l { get; set; } //low
+        public double n { get; set; } // (!)investigate
         public decimal t { get; set; } //unix timestamp
         public long v { get; set; } //volume
